+++ pkg/identity/role.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package util
+package identity
 
 // Role is the operator role, mainly used for logging at the moment.
 type Role int
