+++ pkg/txnutil/lock_resolver.go
@@ -21,7 +21,7 @@ import (
 	"github.com/pingcap/kvproto/pkg/kvrpcpb"
 	"github.com/pingcap/log"
 	"github.com/pingcap/tiflow/cdc/model"
-	"github.com/pingcap/tiflow/pkg/util"
+	"github.com/pingcap/tiflow/pkg/identity"
 	tikverr "github.com/tikv/client-go/v2/error"
 	"github.com/tikv/client-go/v2/tikv"
 	"github.com/tikv/client-go/v2/tikvrpc"
@@ -37,12 +37,12 @@ type LockResolver interface {
 type resolver struct {
 	kvStorage  tikv.Storage
 	changefeed model.ChangeFeedID
-	role       util.Role
+	role       identity.Role
 }
 
 // NewLockerResolver returns a LockResolver.
 func NewLockerResolver(kvStorage tikv.Storage,
-	id model.ChangeFeedID, role util.Role) LockResolver {
+	id model.ChangeFeedID, role identity.Role) LockResolver {
 	return &resolver{
 		kvStorage:  kvStorage,
 		changefeed: id,
